Fix: really install extjs in CDH images at build time
Use the same logic as the configure_extjs snipped used for mapr
(we should have a common file for all of them, with just different
parameters).
It is worth noting that the file is downloaded and copied at runtime
if not found, but better support the offline case.

Story: 2001992
Task: 19618
Change-Id: I5bdded236abd218891cda1bf7d3f25db90c77fdf