Reduce the dependencies, add more common Zuul jobs
The requirements files have been mostly copied from sahara.git,
but most of the dependencies are not needed.

Refresh the content of lower-constraints.txt and ensure that
the lower-constraints tox environment considers also the doc
dependencies.

Explicitly depend on the pre-release of Sahara for Stein.

Add few more jobs to avoid breakages, namely:
- add check-requirements and openstack-lower-constraints-jobs;
- replace tempest-plugin-jobs with openstack-python-jobs,
  openstack-python35-jobs and openstack-python36-jobs
  now that it is possible to depend on sahara.

Change-Id: I087151df07044d9012a6a89bab9925c194b622cd