Limit envlist to py37 for Python 3 Train goal
This is a mechanically generated patch to ensure unit testing is in place
for all of the Tested Runtimes for Train. The correct zuul jobs are
already in place so this change just limits default 'tox' calls to py37
to save developers time.

See the Train python3-updates goal document for details:
https://governance.openstack.org/tc/goals/train/python3-updates.html

Change-Id: I54ff9e8730cb163482b1bcc7983852fe64f61b95
Story: #2005924
Task: #34244