Drop lower-constraints.txt and its testing
As discussed in TC PTG[1] and TC resolution[2], dropping
the lower-constraints.txt file and its testing.
We will keep lower bounds in the requirements.txt file but
with a note that these are not tested lower bounds and we
try our best to keep them updated.

And switch to openstack-python3-zed-jobs

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L326
[2] https://governance.openstack.org/tc/resolutions/20220414-drop-lower-constraints.html#proposal

Change-Id: If1f5aea0975e05b00e57f03d0f61f6f0fa7796b0